refactor: 예약/대기 조회 서비스, 관리 서비스 클래스 분리
