+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -1,6 +1,7 @@
 package roomescape.reservation.presentation;
 
 import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,10 +12,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.member.domain.Member;
-import roomescape.reservation.application.ReservationService;
+import roomescape.reservation.application.ReservationManageService;
+import roomescape.reservation.application.BookingQueryService;
 import roomescape.reservation.application.ReservationTimeService;
 import roomescape.reservation.application.ThemeService;
-import roomescape.reservation.application.WaitingReservationService;
+import roomescape.reservation.application.WaitingQueryService;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
@@ -28,17 +30,23 @@ import java.util.List;
 @RestController
 @RequestMapping("/reservations")
 public class ReservationController {
-    private final ReservationService reservationService;
-    private final WaitingReservationService waitingReservationService;
+    private final BookingQueryService bookingQueryService;
+    private final ReservationManageService waitingScheduler;
+    private final ReservationManageService bookingScheduler;
+    private final WaitingQueryService waitingQueryService;
     private final ReservationTimeService reservationTimeService;
     private final ThemeService themeService;
 
-    public ReservationController(ReservationService reservationService,
-                                 WaitingReservationService waitingReservationService,
+    public ReservationController(BookingQueryService bookingQueryService,
+                                 @Qualifier("waitingManageService") ReservationManageService waitingScheduler,
+                                 @Qualifier("bookingManageService") ReservationManageService bookingScheduler,
+                                 WaitingQueryService waitingQueryService,
                                  ReservationTimeService reservationTimeService,
                                  ThemeService themeService) {
-        this.reservationService = reservationService;
-        this.waitingReservationService = waitingReservationService;
+        this.bookingQueryService = bookingQueryService;
+        this.waitingScheduler = waitingScheduler;
+        this.bookingScheduler = bookingScheduler;
+        this.waitingQueryService = waitingQueryService;
         this.reservationTimeService = reservationTimeService;
         this.themeService = themeService;
     }
@@ -46,21 +54,30 @@ public class ReservationController {
     @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(@RequestBody @Valid ReservationSaveRequest request,
                                                                  Member loginMember) {
+        Reservation newReservation = toNewReservation(request, loginMember);
+        if (newReservation.isBooking()) {
+            Reservation createdReservation = bookingScheduler.create(newReservation);
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body(ReservationResponse.from(createdReservation));
+        }
+        Reservation createdReservation = waitingScheduler.create(newReservation);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(ReservationResponse.from(createdReservation));
+    }
+
+    private Reservation toNewReservation(ReservationSaveRequest request, Member loginMember) {
         ReservationTime reservationTime = reservationTimeService.findById(request.timeId());
         Theme theme = themeService.findById(request.themeId());
-        Reservation newReservation = request.toModel(theme, reservationTime, loginMember);
-        Reservation createReservation = reservationService.create(newReservation);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(ReservationResponse.from(createReservation));
+        return request.toModel(theme, reservationTime, loginMember);
     }
 
     @GetMapping("/mine")
     public ResponseEntity<List<MyReservationResponse>> findMyReservations(Member loginMember) {
-        List<MyReservationResponse> myBookingResponses = reservationService.findReservationsInBookingByMember(loginMember)
+        List<MyReservationResponse> myBookingResponses = bookingQueryService.findReservationsInBookingByMember(loginMember)
                 .stream()
                 .map(MyReservationResponse::from)
                 .toList();
-        List<MyReservationResponse> myWaitingResponses = waitingReservationService.findWaitingReservationsWithPreviousCountByMember(loginMember)
+        List<MyReservationResponse> myWaitingResponses = waitingQueryService.findWaitingReservationsWithPreviousCountByMember(loginMember)
                 .stream()
                 .map(MyReservationResponse::from)
                 .toList();
@@ -71,7 +88,7 @@ public class ReservationController {
 
     @DeleteMapping("/{id}/waiting")
     public ResponseEntity<Void> deleteMyWaitingReservation(@PathVariable Long id, Member loginMember) {
-        waitingReservationService.deleteWaitingReservation(id, loginMember);
+        waitingScheduler.delete(id, loginMember);
         return ResponseEntity.noContent().build();
     }
 }
