+++ src/test/java/roomescape/reservation/application/ReservationTimeServiceTest.java
@@ -3,6 +3,7 @@ package roomescape.reservation.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import roomescape.common.ServiceTest;
 import roomescape.global.exception.ViolationException;
 import roomescape.member.application.MemberService;
@@ -31,7 +32,8 @@ class ReservationTimeServiceTest extends ServiceTest {
     private ReservationTimeService reservationTimeService;
 
     @Autowired
-    private ReservationService reservationService;
+    @Qualifier("bookingManageService")
+    private ReservationManageService bookingManageService;
 
     @Autowired
     private ThemeService themeService;
@@ -85,7 +87,7 @@ class ReservationTimeServiceTest extends ServiceTest {
         ReservationTime reservationTime = reservationTimeService.create(new ReservationTime(MIA_RESERVATION_TIME));
         Theme theme = themeService.create(WOOTECO_THEME());
         Member member = memberService.create(USER_MIA());
-        reservationService.create(MIA_RESERVATION(reservationTime, theme, member, BOOKING));
+        bookingManageService.create(MIA_RESERVATION(reservationTime, theme, member, BOOKING));
 
         // when & then
         assertThatThrownBy(() -> reservationTimeService.delete(reservationTime.getId()))
@@ -99,7 +101,7 @@ class ReservationTimeServiceTest extends ServiceTest {
         ReservationTime miaReservationTime = reservationTimeService.create(new ReservationTime(MIA_RESERVATION_TIME));
         Theme theme = themeService.create(WOOTECO_THEME());
         Member mia = memberService.create(USER_MIA());
-        Reservation miaReservation = reservationService.create(MIA_RESERVATION(miaReservationTime, theme, mia, BOOKING));
+        Reservation miaReservation = bookingManageService.create(MIA_RESERVATION(miaReservationTime, theme, mia, BOOKING));
 
         reservationTimeService.create(new ReservationTime(LocalTime.of(16, 0)));
